Add DispatchingPropertyAnnotator
